Reorganize all agents from agent_classes.py
