+++ pommerman/__init__.py
@@ -5,7 +5,6 @@ import docker
 from . import configs
 from . import utility
 from . import agents
-from . import agent_classes
 
 client = docker.from_env()
 servers = os.environ.get('PLAYGROUND_BATTLE_SERVERS', ','.join(['http://localhost']*4)).split(',')
@@ -32,12 +31,12 @@ def make_agent(config_string, agent_string, agent_id=-1, docker_env_dict=None):
     if agent_type == "player":
         assert agent_control in ["arrows"]
         on_key_press, on_key_release = utility.get_key_control(agent_control)
-        agent_instance = agent_classes.PlayerAgent(
+        agent_instance = agents.PlayerAgent(
             agent, utility.KEY_INPUT, on_key_press=on_key_press, on_key_release=on_key_release)
     elif agent_type == "random":
-        agent_instance = agent_classes.RandomAgent(agent)
+        agent_instance = agents.RandomAgent(agent)
     elif agent_type == "docker":
-        agent_instance = agent_classes.DockerAgent(
+        agent_instance = agents.DockerAgent(
             agent,
             docker_image=agent_control,
             docker_client=client,
