+++ pommerman/agents/__init__.py
@@ -1,1 +1,6 @@
+from .base_agent import BaseAgent
+from .docker_agent import DockerAgent
+from .player_agent import PlayerAgent
+from .random_agent import RandomAgent
 from .simple_agent import SimpleAgent
+from .tensorforce_agent import TensorForceAgent
