+++ pommerman/agents/random_agent.py
@@ -1,0 +1,7 @@
+from . import BaseAgent
+
+
+class RandomAgent(BaseAgent):
+    """The Random Agent that returns random actions given an action_space."""
+    def act(self, obs, action_space):
+        return action_space.sample()
